A couple things: - Few changes (again) on the smeltery structure. - Added missing unfired blowpipe. - Better potash recipe. - Reformated code.

+++ src/main/java/tfctech/objects/storage/MachineEnergyContainer.java
@@ -156,7 +156,7 @@ public class MachineEnergyContainer extends EnergyStorage implements INBTSeriali
         @Override
         public long getInputVoltage()
         {
-            return (long) TechConfig.DEVICES.gtceVoltage;
+            return TechConfig.DEVICES.gtceVoltage;
         }
 
         @Override
